Move ExtendedPortfoliOverview where it is needed
