+++ robozonky-api/src/main/java/com/github/robozonky/internal/tenant/RemotePortfolio.java
@@ -20,7 +20,7 @@ import java.math.BigDecimal;
 import java.util.Map;
 
 import com.github.robozonky.api.remote.enums.Rating;
-import com.github.robozonky.api.strategies.ExtendedPortfolioOverview;
+import com.github.robozonky.api.strategies.PortfolioOverview;
 
 public interface RemotePortfolio {
 
@@ -37,15 +37,9 @@ public interface RemotePortfolio {
 
     Map<Rating, BigDecimal> getTotal();
 
-    Map<Rating, BigDecimal> getAtRisk();
-
-    Map<Rating, BigDecimal> getSellable();
-
-    Map<Rating, BigDecimal> getSellableWithoutFee();
-
     /**
-     * Takes {@link #getBalance()}, {@link #getTotal()} and {@link #getAtRisk()} and summarizes them.
+     * Takes {@link #getBalance()}, and {@link #getTotal()} and summarizes them.
      * @return
      */
-    ExtendedPortfolioOverview getOverview();
+    PortfolioOverview getOverview();
 }
