+++ robozonky-app/src/main/java/com/github/robozonky/app/tenant/RemotePortfolioImpl.java
@@ -18,10 +18,6 @@ package com.github.robozonky.app.tenant;
 
 import java.math.BigDecimal;
 import java.time.Duration;
-import java.time.LocalTime;
-import java.time.OffsetDateTime;
-import java.time.ZoneId;
-import java.time.ZonedDateTime;
 import java.util.Collections;
 import java.util.EnumMap;
 import java.util.LinkedHashMap;
@@ -33,13 +29,11 @@ import java.util.stream.Stream;
 import com.github.robozonky.api.remote.entities.OverallPortfolio;
 import com.github.robozonky.api.remote.entities.RiskPortfolio;
 import com.github.robozonky.api.remote.enums.Rating;
-import com.github.robozonky.api.strategies.ExtendedPortfolioOverview;
+import com.github.robozonky.api.strategies.PortfolioOverview;
 import com.github.robozonky.internal.async.Reloadable;
 import com.github.robozonky.internal.tenant.RemotePortfolio;
 import com.github.robozonky.internal.tenant.Tenant;
-import com.github.robozonky.internal.test.DateUtil;
 import com.github.robozonky.internal.util.BigDecimalCalculator;
-import io.vavr.Tuple2;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -48,11 +42,9 @@ class RemotePortfolioImpl implements RemotePortfolio {
     static final Duration SELLABLE_REFRESH = Duration.ofHours(1);
     private static final Logger LOGGER = LogManager.getLogger(RemotePortfolioImpl.class);
     private final Reloadable<RemoteData> portfolio;
-    private final Reloadable<Map<Rating, BigDecimal>> atRisk;
-    private final Reloadable<Tuple2<Map<Rating, BigDecimal>, Map<Rating, BigDecimal>>> sellable;
     private final AtomicReference<Map<Integer, Blocked>> syntheticByLoanId =
             new AtomicReference<>(new LinkedHashMap<>(0));
-    private final Reloadable<ExtendedPortfolioOverview> portfolioOverview;
+    private final Reloadable<PortfolioOverview> portfolioOverview;
     private final AtomicReference<BigDecimal> lastKnownBalance = new AtomicReference<>(BigDecimal.ZERO);
     private final boolean isDryRun;
 
@@ -62,58 +54,10 @@ class RemotePortfolioImpl implements RemotePortfolio {
                 .reloadAfter(Duration.ofMinutes(5))
                 .finishWith(this::refreshPortfolio)
                 .build();
-        this.atRisk = Reloadable.with(() -> Util.getAmountsAtRisk(tenant))
-                .reloadAfter(Duration.ofMinutes(30)) // not so important, may have a bit of delay
-                .finishWith(this::refreshRisk)
-                .async()
-                .build();
-        this.sellable = Reloadable.with(() -> Util.getAmountsSellable(tenant))
-                .reloadAfter(RemotePortfolioImpl::getSellableRefresh)
-                .finishWith(this::refreshSellable)
-                .async()
-                .build();
-        this.portfolioOverview = Reloadable.with(() -> new PortfolioOverviewImpl(this)
-                .extend(getAtRisk(), getSellable(), getSellableWithoutFee()))
+        this.portfolioOverview = Reloadable.with(() -> (PortfolioOverview) new PortfolioOverviewImpl(this))
                 .finishWith(po -> LOGGER.debug("New portfolio overview: {}.", po))
                 .reloadAfter(Duration.ofMinutes(5))
                 .build();
-    }
-
-    /**
-     * Zonky portfolio refresh usually runs at midnight and at 9am. During that time, Zonky portfolio will report
-     * zero sellable participations, as participation sale is forbidden at the time. Therefore we will avoid updating
-     * sellables during that time.
-     * @param old
-     * @return Duration from now until such a time when the sellable info needs to be refreshed.
-     */
-    static Duration getSellableRefresh(final Tuple2<Map<Rating, BigDecimal>, Map<Rating, BigDecimal>> old) {
-        final OffsetDateTime now = DateUtil.offsetNow();
-        final ZoneId zonkyZone = ZoneId.of("Europe/Prague");
-        final ZonedDateTime zonkyServerTimeWhenNextRefresh = now.atZoneSameInstant(zonkyZone)
-                .plus(SELLABLE_REFRESH);
-        final int refreshHour = zonkyServerTimeWhenNextRefresh.getHour();
-        if (refreshHour > 22) { // ignore 23:00-23:59
-            LOGGER.debug("Sellable refresh will wait until 2am the next day.");
-            final ZonedDateTime target = LocalTime.of(2, 00)
-                    .atDate(now.toLocalDate().plusDays(1))
-                    .atZone(zonkyZone);
-            return Duration.between(now, target);
-        } else if (refreshHour < 2) { // ignore 0:00-02:00
-            LOGGER.debug("Sellable refresh will wait until 2am.");
-            final ZonedDateTime target = LocalTime.of(2, 00)
-                    .atDate(now.toLocalDate())
-                    .atZone(zonkyZone);
-            return Duration.between(now, target);
-        } else if (refreshHour > 7 && refreshHour < 11) { // ignore 8:00-11:00
-            LOGGER.debug("Sellable refresh will wait until 11am.");
-            final ZonedDateTime target = LocalTime.of(11, 00)
-                    .atDate(now.toLocalDate())
-                    .atZone(zonkyZone);
-            return Duration.between(now, target);
-        } else { // check an hour from now
-            LOGGER.debug("Sellable refresh will wait for {}.", SELLABLE_REFRESH);
-            return Duration.between(now, zonkyServerTimeWhenNextRefresh);
-        }
     }
 
     private static BigDecimal sum(final OverallPortfolio portfolio) {
@@ -133,18 +77,6 @@ class RemotePortfolioImpl implements RemotePortfolio {
                 .collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue)));
         LOGGER.debug("New synthetics: {}.", updatedSynthetics);
         // force overview recalculation now that we have the latest portfolio
-        portfolioOverview.clear();
-    }
-
-    private void refreshRisk(final Map<Rating, BigDecimal> data) {
-        LOGGER.debug("New risk data: {}.", data);
-        // force overview recalculation now that we have the latest data
-        portfolioOverview.clear();
-    }
-
-    private void refreshSellable(final Tuple2<Map<Rating, BigDecimal>, Map<Rating, BigDecimal>> data) {
-        LOGGER.debug("New sellability data: {}.", data);
-        // force overview recalculation now that we have the latest data
         portfolioOverview.clear();
     }
 
@@ -207,28 +139,8 @@ class RemotePortfolioImpl implements RemotePortfolio {
     }
 
     @Override
-    public Map<Rating, BigDecimal> getAtRisk() {
-        return atRisk.get().getOrElseThrow(t -> new IllegalStateException("Failed fetching remote risk data.", t));
-    }
-
-    @Override
-    public ExtendedPortfolioOverview getOverview() {
+    public PortfolioOverview getOverview() {
         return portfolioOverview.get()
                 .getOrElseThrow(t -> new IllegalStateException("Failed calculating portfolio overview.", t));
-    }
-
-    private Tuple2<Map<Rating, BigDecimal>, Map<Rating, BigDecimal>> getSellabilityData() {
-        return sellable.get()
-                .getOrElseThrow(t -> new IllegalStateException("Failed fetching remote sellability data.", t));
-    }
-
-    @Override
-    public Map<Rating, BigDecimal> getSellable() {
-        return getSellabilityData()._1();
-    }
-
-    @Override
-    public Map<Rating, BigDecimal> getSellableWithoutFee() {
-        return getSellabilityData()._2();
     }
 }
