+++ robozonky-app/src/test/java/com/github/robozonky/app/tenant/UtilTest.java
@@ -17,22 +17,15 @@
 package com.github.robozonky.app.tenant;
 
 import java.math.BigDecimal;
-import java.util.Collections;
 import java.util.Map;
 import java.util.stream.Stream;
 import javax.ws.rs.NotFoundException;
 
-import com.github.robozonky.api.remote.entities.CurrentOverview;
-import com.github.robozonky.api.remote.entities.RiskPortfolio;
-import com.github.robozonky.api.remote.entities.Statistics;
-import com.github.robozonky.api.remote.entities.sanitized.Investment;
 import com.github.robozonky.api.remote.entities.sanitized.Loan;
 import com.github.robozonky.api.remote.enums.Rating;
 import com.github.robozonky.app.AbstractZonkyLeveragingTest;
-import com.github.robozonky.internal.remote.Select;
 import com.github.robozonky.internal.remote.Zonky;
 import com.github.robozonky.internal.tenant.Tenant;
-import io.vavr.Tuple2;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.*;
@@ -42,48 +35,7 @@ import static org.mockito.Mockito.*;
 
 class UtilTest extends AbstractZonkyLeveragingTest {
 
-    @Test
-    void atRisk() {
-        final Investment i = Investment.custom()
-                .setRating(Rating.D)
-                .setRemainingPrincipal(BigDecimal.TEN)
-                .setPaidInterest(BigDecimal.ZERO)
-                .setPaidPenalty(BigDecimal.ZERO)
-                .build();
-        final Statistics stats = mock(Statistics.class);
-        final RiskPortfolio r = new RiskPortfolio(i.getRating(), 0, 0, i.getRemainingPrincipal().longValue());
-        when(stats.getRiskPortfolio()).thenReturn(Collections.singletonList(r));
-        when(stats.getCurrentOverview()).thenReturn(mock(CurrentOverview.class));
-        final Zonky zonky = harmlessZonky(10_000);
-        when(zonky.getStatistics()).thenReturn(stats);
-        when(zonky.getDelinquentInvestments()).thenReturn(Stream.of(i));
-        final Tenant tenant = mockTenant(zonky);
-        final Map<Rating, BigDecimal> result = Util.getAmountsAtRisk(tenant);
-        assertThat(result).containsOnlyKeys(Rating.D);
-        assertThat(result.get(Rating.D).longValue()).isEqualTo(BigDecimal.TEN.longValue());
-    }
-
-    @Test
-    void sellable() {
-        final Investment i = Investment.custom()
-                .setRating(Rating.D)
-                .setRemainingPrincipal(BigDecimal.TEN)
-                .setSmpFee(BigDecimal.ONE)
-                .build();
-        final Investment i2 = Investment.custom()
-                .setRating(Rating.A)
-                .setRemainingPrincipal(BigDecimal.ONE)
-                .setSmpFee(BigDecimal.ZERO)
-                .build();
-        final Zonky zonky = harmlessZonky(10_000);
-        when(zonky.getInvestments((Select)any())).thenReturn(Stream.of(i, i2));
-        final Tenant tenant = mockTenant(zonky);
-        final Tuple2<Map<Rating, BigDecimal>, Map<Rating, BigDecimal>> result = Util.getAmountsSellable(tenant);
-        assertThat(result._1).containsOnlyKeys(Rating.D, Rating.A);
-        assertThat(result._2).containsOnlyKeys(Rating.A);
-    }
-
-    @Test
+   @Test
     void ignoresBlockedAmountsFromZonky() {
         final Zonky zonky = harmlessZonky(10_000);
         final Tenant tenant = mockTenant(zonky);
